refactored StringAssertion to be separate from QuantityAssertion
This is a breaking change since the methods from QuantityAssertion are not available to StringAssertion anymore. An extra layer has been added with ObjectAssertion to capture methods that were used in both interfaces.
